Refactor packages and healthcheck
