+++ src/main/java/uk/gov/companieshouse/ocrapiconsumer/request/ExtractTextResultDTO.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.ocrapiconsumer.request.extractedtext;
+package uk.gov.companieshouse.ocrapiconsumer.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
