+++ src/test/java/uk/gov/companieshouse/ocrapiconsumer/request/TestParent.java
@@ -1,7 +1,6 @@
 package uk.gov.companieshouse.ocrapiconsumer.request;
 
 import org.springframework.http.ResponseEntity;
-import uk.gov.companieshouse.ocrapiconsumer.request.extractedtext.ExtractTextResultDTO;
 
 public class TestParent {
     protected static final String EXTERNAL_REFERENCE_ID = "ABC";
